feat(settings): move LLM/TTS/STT settings into sidebar pages
Moves AI service configuration out of the Character page into dedicated
settings pages accessible from the settings sidebar:

- Added Settings > LLM Model page
- Added Settings > TTS page
- Added Settings > STT page
- Extracted reusable LlmSettings/TtsSettings/SttSettings components
- Extracted shared ai-services-settings state and testable logic helpers
- Removed the AI Services collapsible section from Character settings
- Updated README configuration instructions

No new providers or settings fields were added; existing options were
moved as-is.
