+++ src/routes/app/settings/+layout.svelte
@@ -8,6 +8,9 @@
 	const navItems = $derived([
 		{ href: localPath('app', '/settings/persona'), label: 'Character', icon: 'persona' },
 		{ href: localPath('app', '/settings/display'), label: 'Display', icon: 'monitor' },
+		{ href: localPath('app', '/settings/llm'), label: 'LLM Model', icon: 'brain' },
+		{ href: localPath('app', '/settings/tts'), label: 'TTS', icon: 'mic' },
+		{ href: localPath('app', '/settings/stt'), label: 'STT', icon: 'mic' },
 		{ href: localPath('app', '/settings/data'), label: 'Data', icon: 'database' },
 		{ href: localPath('app', '/settings/developer'), label: 'Developer', icon: 'code' }
 	]);
